Consolidate *.Fast.cs
System.Memory OOB package that this split was originally done for does not exist anymore.
